refactor : CircleIconWrapper import
